Add a new boolean property and a new method!
The boolean property says if the Dali Signal is inverted or not.
And the new methods are about starting and stopping the Idle state,
the TX pin is switched to an input for starting the idle state and is
switche d to an output if the sending starts. So the Idke state is
undepndent of the Dali signals are inverted or not.